requirements.txt
aniso8601==8.0.0
appdirs==1.4.3
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.8.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
colorama==0.4.3
configparser==5.0.0
databricks-cli==0.10.0
databricks-connect==6.4.0
distlib==0.3.0
dnspython==1.16.0
ez-setup==0.9
filelock==3.0.12
Flask==1.1.2
Flask-Babel==1.0.0
Flask-Cors==3.0.4
Flask-Mail==0.9.1
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
Flask-Table==0.5.0
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
mysql-connector-python==8.0.19
mysqlclient==1.4.6
numpy==1.18.3
packaging==20.3
pandas==1.0.3
pandasql==0.7.3
pluggy==0.13.1
protobuf==3.6.1
py==1.8.1
py4j==0.10.7
pylint==2.4.4
PyMySQL==0.9.3
pyparsing==2.4.7
pytest==5.4.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.16
tabulate==0.8.7
urllib3==1.25.9
virtualenv==20.0.20
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.11.2
